Split sensor.py into modules, fix band-formatting bug, log bloat, missing dep
Repo-wide cleanup pass, verified with a real import-resolution test
(mocked HA modules, actual module loading -- not just py_compile syntax
checks) before deploying.

Split sensor.py (1576 lines, mixing several unrelated concerns) into:
- sensor_base.py: ZTERouterEntity base class + guard_stale_data decorator
  (shared by sensor.py, sensor_bands.py, sensor_flux.py)
- coordinators.py: ZTERouterDataUpdateCoordinator,
  ZTERouterSMSUpdateCoordinator, extract_json
- sensor_bands.py: format_ca_bands, derive_primary_band_from_lteca,
  calculate_enodeb_id, ConnectedBandsSensor
- sensor_flux.py: FLUX sensors (ZTEDataStatisticsSensor, ZTEFluxSensor,
  ZTEFluxTotalUsageSensor) and their BYTE_KEYS/THROUGHPUT_KEYS/TEXT_KEYS
- sensor.py: everything else (generic sensor, SMS/client/usage sensors,
  async_setup_entry), importing from the above
__init__.py now imports the coordinators from .coordinators instead of
.sensor (the only external call site).

Fixed a real, live-confirmed bug while splitting out sensor_bands.py:
format_ca_bands() and derive_primary_band_from_lteca() both read
band_info[0] as the band number for 5-field CA records, but that
position is actually the PCI (format is pci,band,cellid,arfcn,
bandwidth, confirmed against live G5 Ultra data). This is exactly why
the "Connected Bands" sensor showed "B281@20.0MHz" three times in a
row earlier this session -- 281 was the PCI, not three carriers all on
the same impossible band. Fixed to read band_info[1]; verified against
the real captured data that it now produces the correct B3+B1+B20.

Also, from the broader repo review:
- manifest.json was missing "requests" from requirements despite
  g5_ultra_client.py importing it directly -- worked by luck as a
  transitive dependency on most installs, would ModuleNotFoundError on
  a minimal one. Added.
- g5_ultra_client.py maintained its own DEBUG-forced log file
  (ultra.log) written into the integration's own folder, bypassing
  Home Assistant's logging config entirely -- this is why it grew to
  199 MB during this session. Removed the custom handler; now uses
  HA's standard logger (respects the configured level, uses HA's own
  log rotation).
- mc.py has the same DEBUG-forced pattern in its __main__ block, which
  runs on every single MC-series command since it's invoked as a
  subprocess per call. Can't simply defer to HA's logging (separate
  process), so lowered the default to WARNING instead to cut volume
  while keeping mc.log useful for troubleshooting failures.
- hacs.json had no minimum Home Assistant version pin; added one so
  HACS itself can catch "installed on too-old HA" before a confusing
  runtime error.

manifest.json: 1.0.56b8.
