+++ hacs.json
@@ -2,5 +2,6 @@
   "name": "ZTE Router",
   "render_readme": true,
   "country": "HR",
-	"hide_default_branch": true
+  "hide_default_branch": true,
+  "homeassistant": "2023.1.0"
 }
